backend/requirements.txt
Flask==3.0.3
gunicorn==22.0.0
requests==2.32.3
python-dotenv==1.0.1

# PDF generation and inspection
reportlab==4.2.5
pikepdf==9.4.0
pdfplumber==0.11.4

# Document parsing for reference extraction
python-docx==1.1.2
openpyxl==3.1.5
